requirements/base.txt
# https://docs.djangoproject.com/en/1.7/
Django==1.7.2

# Configuration
django-secure==1.0.1
dj-database-url==0.3.0

# PostgreSQL adapter
# http://pythonhosted.org/psycopg2/
psycopg2==2.5.4

# Lots of custom mixins
# http://django-braces.readthedocs.org/en/v1.4.0/
django-braces==1.4.0

# Cool class-based views
django-extra-views==0.6.5

# Control of output of form rendering
# http://django-floppyforms.readthedocs.org/en/latest/
django-floppyforms==1.3.0

# Serve production static files with Django
# https://github.com/kennethreitz/dj-static
dj-static==0.0.6

# Useful model mixins and utilities such as `TimeStampedModel` and `Choices`
# https://django-model-utils.readthedocs.org/en/latest/
django-model-utils==2.2

# CSS and JS compressor and compiler. Also minifies HTML
# http://django-pipeline.readthedocs.org/en/latest/
django-pipeline==1.4.2

# Facebook registration/authorization
django-allauth==0.19.0

django-phonenumber-field==0.6

# Memcached
python-memcached==1.53

Pillow==2.7.0

# Form fields styling directly in templates
django-widget-tweaks==1.3

django-aggregate-if==0.5

# Translation library
django-rosetta==0.7.4

# Image crop lib
django-image-cropping==1.0.1
easy-thumbnails==2.2

# Rich text editor
django-ckeditor==4.4.7

# Translation library
django-hvad==1.2.0

# solr python adapter
pysolr==3.3.0

# Django search platform
django-haystack==2.3.1

# Django slugify application that guarantees Uniqueness and handles Unicode
django-uuslug==1.0.3

# Admin sortable
django-admin-sortable2==0.3.3

# SKD tools to check file types
django-skd-tools[TypedFileField]==0.2.2

# country codes lib with CountryField
django-countries==3.3

# http library
requests==2.6.0

# ip retriever
django-ipware==1.0.0

pytz

# async task lib
celery==3.1.18

# adapter
django-celery==3.1.16

# celery email backend
django-celery-email==1.1.1

django-multiselectfield==0.1.3
